Test certmonger-generated certificates directly in cockpit/ws-certs.d/
Change tests_certificate.yml to a cleaner approach which directly asks
certmonger to put the certificate into /etc/cockpit/ws-certs.d/, and
avoid the `runafter` script.  This approach relies on cockpit ≥ 211 and
https://github.com/linux-system-roles/certificate/pull/97,
so it does not work on RHEL/CentOS 7. Skip it with older cockpit
versions.

Move the previous test to tests_certificate_runafter.yml, to cover the
`runafter` script approach which works everywhere.