Test certmonger-generated certificates directly in cockpit/ws-certs.d/ (#40)
* Add names to all tasks in tests_certificate.yml

Required by ansible-lint 2.11.

* tests: Check for existing roles/linux-system-roles.certificate dir only

The internal downstream tests seem to check out the git tree in a way
that does not create .git/HEAD. Drop this assumption, and test for the
directory only.

* Test certmonger-generated certificates directly in cockpit/ws-certs.d/

Change tests_certificate.yml to a cleaner approach which directly asks
certmonger to put the certificate into /etc/cockpit/ws-certs.d/, and
avoid the `runafter` script.  This approach relies on cockpit ≥ 211 and
https://github.com/linux-system-roles/certificate/pull/97,
so it does not work on RHEL/CentOS 7. Skip it with older cockpit
versions.

Move the previous test to tests_certificate_runafter.yml, to cover the
`runafter` script approach which works everywhere.