tests: Generate self-signed certificate for RHEL >= 8 (#37)
`ca: self-sign` is only broken on RHEL/CentOS 7, but works everywhere
else. So let's use it in tests_certificate.yml to cover this case as
well, and make sure it does not break again.

tests_certificate_runafter.yml still needs to keep `ca: local` and the
additional CA retrieval step, but it doesn't hurt to cover that mode.