change recursive role symlink to individual role dir symlinks (#49)
The problem is that recursive symlinks break Ansible Tower/AAP:
https://github.com/linux-system-roles/firewall/issues/60

The network role had already solved this problem, so we just adopted their solution to the other roles.

This may introduce tech debt, as over time there might be obsolete or missing symlinks.  It should be very rare to add a new role directory to a role. But if it becomes a problem, we can add a git commit hook or something like that.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>