fix: wildcard package installation not working with dnf module
Cause: The dnf module is not installing cockpit-*

Consequence: The full list of requested packages is not installed.

Fix: Call the dnf command directly with the wildcard package name
and the packages to exclude.

Result: The requested wildcard packages are installed.

This also adds a check to the test to ensure the excluded
packages are not installed.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>