fix: Reset certificate facts between playbook runs
When running `tests_certificate_external.yml tests/tests_default.yml`
tox tests, the first test left non-empty `cockpit_{cert,private_key}`
facts in Ansible's brain. This broke the second run in
tests_default.yml, as that then created symlinks to files which no
longer exist (as the actual certificates get removed in
tests/tasks/cleanup.yml), and thus the "cockpit works with TLS" step
failed as cockpit.service kept failing.

There is no way to unset a fact (setting it to `${{ undef() }}` errors
out), so set them to an empty value, and replace "is defined" with an
"is nonempty" check.