fix: Skip runtime operations in non-systemd environments
Cause: Both the actual role and the tests were previously assuming that
the system was booted with systemd and could run services. But this is
not the case when running the role during container builds.

Consequence: The role did not work during bootc container builds.

Fix: Detect if the system is booted (with systemd), and skip all runtime
operations and checks if not.

Result: The role now works during container builds.

Add a config file only test to tests_port.yml to cover container runs,
as all the other assertions depend on cockpit running. Skip the
certmonger tests for non-booted environments, as certmonger requires a
running system by design (these are a case for running the role in a
deployed system). Also skip the firewall managed related checks for the
time being, until the firewall role works during container builds
(https://issues.redhat.com/browse/RHEL-88425).

https://issues.redhat.com/browse/RHEL-88423