feat: Support this role in container builds
Feature: Support running the cockpit role during container builds.

Reason: This is particularly useful for building bootc derivative OSes.

Result: These flags enable running the bootc container scenarios in CI,
which ensures that the role works in buildah build environment. This
allows us to officially support this role for image mode builds.

Do *not* enable the role for system containers (the `container`) flag.
That currently fails due to SELinux not working properly there, and
needs to be looked at separately if desired.

https://issues.redhat.com/browse/RHEL-88423