ci: Add support for bootc end-to-end validation tests
These tests run the role during a bootc container image build, deploy
the container into a QEMU VM, boot that, and validate the expected
configuration there. They run in two different tox environments, and
thus have to be run in two steps (preparation in buildah, validation in
QEMU). The preparation is expected to output a qcow2 image in
`tests/tmp/TESTNAME/qcow2/disk.qcow2`, i.e. the output structure of
<https://github.com/osbuild/bootc-image-builder>.

There are two possibilities:

 * Have separate bootc end-to-end tests. These are tagged with
   `tests::bootc-e2` and are skipped in the normal qemu-* scenarios.
   They run as part of the container-* ones.

 * Modify an existing test: These need to build a qcow2 image exactly
   *once* (via calling `bootc-buildah-qcow.sh`) and skip setup/cleanup
   and role invocations in validation mode, i.e. when
   `__bootc_validation` is true.

In the container scenario, run the QEMU validation as a separate step in
the workflow.

See https://issues.redhat.com/browse/RHEL-88396